+++ apps/meteor/app/threads/server/functions.ts
@@ -2,7 +2,7 @@ import type { IMessage } from '@rocket.chat/core-typings';
 import { isEditedMessage } from '@rocket.chat/core-typings';
 import { Messages, Subscriptions, ReadReceipts, NotificationQueue } from '@rocket.chat/models';
 
-import { getMentions } from '../../lib/server/lib/notifyUsersOnMessage';
+import { getMentions, getUserIdsFromHighlights } from '../../lib/server/lib/notifyUsersOnMessage';
 
 export async function reply({ tmid }: { tmid?: string }, message: IMessage, parentMessage: IMessage, followers: string[]) {
 	const { rid, ts, u } = message;
@@ -19,7 +19,9 @@ export async function reply({ tmid }: { tmid?: string }, message: IMessage, pare
 			...(Array.isArray(parentMessage.replies) && parentMessage.replies.length ? [u._id] : [parentMessage.u._id, u._id]),
 		]),
 	];
+	const highlightedUserIds = new Set<string>();
 
+	(await getUserIdsFromHighlights(rid, message)).forEach((uid) => highlightedUserIds.add(uid));
 	await Messages.updateRepliesByThreadId(tmid, addToReplies, ts);
 	await ReadReceipts.setAsThreadById(tmid);
 
@@ -35,8 +37,15 @@ export async function reply({ tmid }: { tmid?: string }, message: IMessage, pare
 		await Subscriptions.addUnreadThreadByRoomIdAndUserIds(rid, repliesFiltered, tmid, {});
 	}
 
-	for await (const userId of mentionIds) {
+	const mentionedUsers = new Set<string>([...mentionIds, ...highlightedUserIds]);
+	for await (const userId of mentionedUsers) {
 		await Subscriptions.addUnreadThreadByRoomIdAndUserIds(rid, [userId], tmid, { userMention: true });
+	}
+
+	const highlightIds = Array.from(highlightedUserIds);
+	if (highlightIds.length) {
+		await Subscriptions.setAlertForRoomIdAndUserIds(rid, highlightIds);
+		await Subscriptions.setOpenForRoomIdAndUserIds(rid, highlightIds);
 	}
 }