+++ apps/meteor/client/lib/2fa/utils.ts
@@ -12,6 +12,12 @@ export const isTotpInvalidError = (error: unknown): error is Meteor.Error & ({ e
 	(error as { error?: unknown } | undefined)?.error === 'totp-invalid' ||
 	(error as { errorType?: unknown } | undefined)?.errorType === 'totp-invalid';
 
+export const isTotpMaxAttemptsError = (
+	error: unknown,
+): error is Meteor.Error & ({ error: 'totp-max-attempts' } | { errorType: 'totp-max-attempts' }) =>
+	(error as { error?: unknown } | undefined)?.error === 'totp-max-attempts' ||
+	(error as { errorType?: unknown } | undefined)?.errorType === 'totp-max-attempts';
+
 const isLoginCancelledError = (error: unknown): error is Meteor.Error =>
 	error instanceof Meteor.Error && error.error === Accounts.LoginCancelledError.numericError;