+++ apps/meteor/ee/app/livechat-enterprise/server/hooks/onTransferFailure.ts
@@ -38,35 +38,27 @@ const onTransferFailure = async (
 	}
 
 	// TODO: find enabled not archived here
-	const fallbackDepartment = await LivechatDepartment.findOneById(department.fallbackForwardDepartment, {
-		projection: { name: 1, _id: 1 },
-	});
-
-	if (!fallbackDepartment) {
-		return false;
-	}
-
-	const fallbackDepartmentDb = await LivechatDepartment.findOneById<Pick<ILivechatDepartment, '_id' | 'name'>>(
+	const fallbackDepartment = await LivechatDepartment.findOneById<Pick<ILivechatDepartment, '_id' | 'name'>>(
 		department.fallbackForwardDepartment,
 		{
 			projection: { name: 1, _id: 1 },
 		},
 	);
 
+	if (!fallbackDepartment) {
+		return false;
+	}
+
 	const transferDataFallback = {
 		...transferData,
 		prevDepartment: department.name,
 		departmentId: department.fallbackForwardDepartment,
-		...(fallbackDepartmentDb && {
-			department: fallbackDepartmentDb,
-		}),
+		department: fallbackDepartment,
 	};
 
 	const forwardSuccess = await forwardRoomToDepartment(room, guest, transferDataFallback);
 	if (forwardSuccess) {
 		const { _id, username } = transferData.transferredBy;
-		// The property is injected dynamically on ee folder
-
 		await Message.saveSystemMessage(
 			'livechat_transfer_history_fallback',
 			room._id,