+++ apps/meteor/server/models/raw/Messages.ts
@@ -9,7 +9,6 @@ import type {
 	IMessageWithPendingFileImport,
 } from '@rocket.chat/core-typings';
 import type { FindPaginated, IMessagesModel } from '@rocket.chat/model-typings';
-import { Rooms } from '@rocket.chat/models';
 import type { PaginatedRequest } from '@rocket.chat/rest-typings';
 import { escapeRegExp } from '@rocket.chat/string-helpers';
 import type {
@@ -1340,8 +1339,6 @@ export class MessagesRaw extends BaseRaw<IMessage> implements IMessagesModel {
 
 		const data = Object.assign(record, extraData);
 
-		await Rooms.incMsgCountById(rid, 1);
-
 		return this.insertOne(data);
 	}
 
@@ -1464,10 +1461,6 @@ export class MessagesRaw extends BaseRaw<IMessage> implements IMessagesModel {
 
 		if (!limit) {
 			const count = (await this.deleteMany(query)).deletedCount - notCountedMessages;
-			if (count) {
-				// decrease message count
-				await Rooms.decreaseMessageCountById(rid, count);
-			}
 
 			return count;
 		}
@@ -1480,11 +1473,6 @@ export class MessagesRaw extends BaseRaw<IMessage> implements IMessagesModel {
 					},
 				})
 			).deletedCount - notCountedMessages;
-
-		if (count) {
-			// decrease message count
-			await Rooms.decreaseMessageCountById(rid, count);
-		}
 
 		return count;
 	}